+++ app/models/claim/verify_response_parameters_parser.rb
@@ -1,0 +1,122 @@
+# frozen_string_literal: true
+
+class Claim
+  # Used to convert the parameters from a successful Verify::Response into an
+  # attributes Hash that can be used to update a claim. For example,
+  #
+  #   Claim::VerifyResponseParametersParser.new(verify_response.parameters).attributes
+  #   => { full_name: "Margaret Elaine Heafield Hamilton", date_of_birth: "1936-08-17", ...}
+  #
+  # As well as keys for the personal information of the user (name, address,
+  # etc) the returned Hash includes two additional keys:
+  #
+  #   verified_fields – these are the keys for the attributes that have come
+  #                     back in the Verify response. Recording these allows us
+  #                     to determine which attributes came from Verify and
+  #                     therefore should not be editable by the user.
+  #
+  #   verify_response – this is the entire Verify response payload. We record
+  #                     this so that we can examine the original payload to
+  #                     debug issues in the way we parse responses, but also so
+  #                     we can retrospectively re-parse responses in future if
+  #                     necessary.
+  #
+  class VerifyResponseParametersParser
+    class MissingResponseAttribute < StandardError; end
+
+    def initialize(response_parameters)
+      @response_parameters = response_parameters
+    end
+
+    def attributes
+      identity_attributes.merge(verified_fields: verified_fields, verify_response: @response_parameters)
+    end
+
+    def gender
+      {"MALE" => :male, "FEMALE" => :female}[verify_attributes.dig("gender", "value")]
+    end
+
+    def date_of_birth
+      verify_attributes.fetch("datesOfBirth").first.fetch("value")
+    end
+
+    def first_name
+      most_recent_verified("firstNames")
+    end
+
+    def surname
+      most_recent_verified("surnames")
+    end
+
+    def middle_name
+      most_recent_verified("middleNames", required: false)
+    end
+
+    def full_name
+      [first_name, middle_name, surname].compact.join(" ")
+    end
+
+    def postcode
+      most_recent_address&.fetch("postCode")
+    end
+
+    def address_line_1
+      address_lines[0]
+    end
+
+    def address_line_2
+      address_lines[1]
+    end
+
+    def address_line_3
+      address_lines[2]
+    end
+
+    def address_line_4
+      address_lines[3]
+    end
+
+    private
+
+    def verify_attributes
+      @response_parameters.fetch("attributes")
+    end
+
+    def most_recent_verified(attribute_name, required: true)
+      raise MissingResponseAttribute, "No verified value found for #{attribute_name}" if required && no_verified_values?(attribute_name)
+
+      verify_attributes.fetch(attribute_name)
+        .select { |value| value["verified"] }
+        .max_by { |value| Date.parse(value.fetch("from", Date.today.to_s)) }&.fetch("value")
+    end
+
+    def no_verified_values?(attribute_name)
+      verify_attributes.fetch(attribute_name).none? { |value| value["verified"] }
+    end
+
+    def most_recent_address
+      @most_recent_address ||= most_recent_verified("addresses", required: false)
+    end
+
+    def address_lines
+      most_recent_address&.fetch("lines") || []
+    end
+
+    def identity_attributes
+      @identity_attributes ||= {
+        full_name: full_name,
+        date_of_birth: date_of_birth,
+        payroll_gender: gender,
+        address_line_1: address_line_1,
+        address_line_2: address_line_2,
+        address_line_3: address_line_3,
+        address_line_4: address_line_4,
+        postcode: postcode,
+      }.compact
+    end
+
+    def verified_fields
+      identity_attributes.keys
+    end
+  end
+end