+++ spec/lib/verify/response_spec.rb
@@ -16,7 +16,6 @@ RSpec.describe Verify::Response, type: :model do
 
       expect(verify_response).to be_kind_of(Verify::Response)
       expect(verify_response).to be_verified
-      expect(verify_response.claim_parameters[:full_name]).to eq("Isambard Kingdom Brunel")
     end
   end
 
@@ -25,119 +24,6 @@ RSpec.describe Verify::Response, type: :model do
 
     it "is verified" do
       expect(subject.verified?).to eq(true)
-    end
-
-    it "returns the expected verified parameters" do
-      expect(subject.claim_parameters[:full_name]).to eq("Isambard Kingdom Brunel")
-      expect(subject.claim_parameters[:address_line_1]).to eq("Verified Building")
-      expect(subject.claim_parameters[:address_line_2]).to eq("Verified Street")
-      expect(subject.claim_parameters[:address_line_3]).to eq("Verified Town")
-      expect(subject.claim_parameters[:address_line_4]).to eq("Verified County")
-      expect(subject.claim_parameters[:postcode]).to eq("M12 345")
-      expect(subject.claim_parameters[:date_of_birth]).to eq("1806-04-09")
-      expect(subject.claim_parameters[:payroll_gender]).to eq(:male)
-      expect(subject.claim_parameters[:verified_fields]).to match_array([
-        :full_name,
-        :address_line_1,
-        :address_line_2,
-        :address_line_3,
-        :address_line_4,
-        :postcode,
-        :date_of_birth,
-        :payroll_gender,
-      ])
-      expect(subject.claim_parameters[:verify_response]).to eq(response)
-    end
-
-    context "when the gender from Verify is female" do
-      let(:response_filename) { "identity-verified-female.json" }
-
-      it "returns the expected payroll_gender" do
-        expect(subject.claim_parameters[:payroll_gender]).to eq(:female)
-      end
-
-      it "returns payroll_gender in the verified fields" do
-        expect(subject.claim_parameters[:verified_fields]).to include(:payroll_gender)
-      end
-    end
-
-    context "when the gender from Verify is not specified" do
-      let(:response_filename) { "identity-verified-not-specified-gender.json" }
-
-      it "returns nil for payroll_gender" do
-        expect(subject.claim_parameters[:payroll_gender]).to eq(nil)
-      end
-
-      it "does not return payroll_gender in the verified fields" do
-        expect(subject.claim_parameters[:verified_fields]).to_not include(:payroll_gender)
-      end
-    end
-
-    context "when the gender from Verify is other" do
-      let(:response_filename) { "identity-verified-other-gender.json" }
-
-      it "returns nil for payroll_gender" do
-        expect(subject.claim_parameters[:payroll_gender]).to eq(nil)
-      end
-
-      it "does not return payroll_gender in the verified fields" do
-        expect(subject.claim_parameters[:verified_fields]).to_not include(:payroll_gender)
-      end
-    end
-
-    context "when the address is missing (for EU citizens)" do
-      let(:response_filename) { "identity-verified-no-address.json" }
-
-      it "returns nil for address attributes" do
-        expect(subject.claim_parameters[:address_line_1]).to eq(nil)
-        expect(subject.claim_parameters[:address_line_2]).to eq(nil)
-        expect(subject.claim_parameters[:address_line_3]).to eq(nil)
-        expect(subject.claim_parameters[:postcode]).to eq(nil)
-      end
-
-      it "does not return any address attributes in the verified fields" do
-        expect(subject.claim_parameters[:verified_fields]).to_not include(
-          :address_line_1,
-          :address_line_2,
-          :address_line_3,
-          :postcode
-        )
-      end
-    end
-
-    context "when the middle name isn't verified" do
-      let(:response_filename) { "identity-verified-unverified-middlename.json" }
-
-      it "returns the full name without the unverified middle name" do
-        expect(subject.claim_parameters[:full_name]).to eq("Isambard Brunel")
-      end
-    end
-
-    context "when the last name isn't verified" do
-      let(:response_filename) { "identity-verified-unverified-lastname.json" }
-
-      it "raises an exception" do
-        expect { subject.claim_parameters[:full_name] }.to raise_exception(
-          Verify::Response::MissingResponseAttribute, "No verified value found for surnames"
-        )
-      end
-    end
-  end
-
-  context "with the minimum verified response" do
-    let(:response_filename) { "identity-verified-minimum.json" }
-
-    it "is verified" do
-      expect(subject.verified?).to eq(true)
-    end
-
-    it "returns the expected verified parameters" do
-      expect(subject.claim_parameters[:full_name]).to eq("Isambard Brunel")
-      expect(subject.claim_parameters[:address_line_1]).to eq("Verified Street")
-      expect(subject.claim_parameters[:address_line_2]).to eq("Verified Town")
-      expect(subject.claim_parameters[:address_line_3]).to be_nil
-      expect(subject.claim_parameters[:postcode]).to eq("M12 345")
-      expect(subject.claim_parameters[:date_of_birth]).to eq("1806-04-09")
     end
   end